[tuanNA1] Validate request to system source

+++ eduhub-server/src/main/java/vn/com/eduhub/dto/res/ObjectDataRes.java
@@ -10,6 +10,6 @@ import java.util.List;
 @AllArgsConstructor
 @ToString
 public class ObjectDataRes<T> {
-    private Long totalData;
+    private int totalData;
     private List<T> datas;
 }
